fix(sf2): listEnd type mismatch (Long -> Int coercion)
Kotlin ByteBuffer.position(Int) requires Int; chunkSize is Long because
we read it via 'bb.int.toLong() and 0xFFFFFFFFL'. Adding .toInt() on
the listEnd computation fixes the compile error at SoundFontParser.kt:112.

+++ app/src/main/java/com/t2v/core/midi/sf2/SoundFontParser.kt
@@ -47,7 +47,7 @@ object SoundFontParser {
             val chunkSize = bb.int.toLong() and 0xFFFFFFFFL
             if (chunkId == "LIST") {
                 val listType = readAscii(bb, 4)
-                val listEnd = bb.position() + (chunkSize - 4)
+                val listEnd = (bb.position() + (chunkSize - 4)).toInt()
                 when (listType) {
                     "INFO" -> {
                         // read version (ifil chunk)
